+++ src/dotnetCampus.FileDownloader.WPF/AddDownloadFilePage.xaml
@@ -5,6 +5,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:dotnetCampus.FileDownloader.WPF"
              mc:Ignorable="d" 
+             x:Name="Root"
              d:DesignHeight="450" d:DesignWidth="800">
     <Grid>
         <Grid VerticalAlignment="Center" HorizontalAlignment="Center">
@@ -15,17 +16,20 @@
             </Grid.RowDefinitions>
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="Auto">
-                    
+
                 </ColumnDefinition>
                 <ColumnDefinition Width="*" MinWidth="200"></ColumnDefinition>
             </Grid.ColumnDefinitions>
 
             <TextBlock Text="文件路径：" Margin="10,10,10,10"></TextBlock>
-            <TextBox Grid.Row="0" Grid.Column="1" Margin="10,10,10,10"></TextBox>
+            <TextBox Grid.Row="0" Grid.Column="1" Margin="10,10,10,10"
+                     Text="{Binding ElementName=Root,Path=FilePath}"></TextBox>
             <TextBlock Grid.Row="1" Grid.Column="0" Margin="10,10,10,10" Text="下载链接："></TextBlock>
-            <TextBox Grid.Row="1" Grid.Column="1" Margin="10,10,10,10"></TextBox>
+            <TextBox Grid.Row="1" Grid.Column="1" Margin="10,10,10,10"
+                     Text="{Binding ElementName=Root,Path=Url}"></TextBox>
 
-            <Button Grid.Row="2" Grid.Column="1" Margin="10,5,10,10" HorizontalAlignment="Right" VerticalAlignment="Center" Content="下载"></Button>
+            <Button Grid.Row="2" Grid.Column="1" Margin="10,5,10,10" HorizontalAlignment="Right" VerticalAlignment="Center" Content="下载"
+                    Click="DownloadButton_OnClick"></Button>
         </Grid>
     </Grid>
 </UserControl>
