+++ src/dotnetCampus.FileDownloader.WPF/MainWindow.xaml.cs
@@ -29,18 +29,10 @@ namespace dotnetCampus.FileDownloader.WPF
         {
             
         }
-    }
 
-    public class DownloadFileInfo
-    {
-        public string FileName { get; set; }
-
-        public DateTime AddedTime { get; set; }
-
-        public string FileSize { get; set; }
-
-        public string DownloadProcess { get; set; }
-
-        public string DownloadSpeed { get; set; }
+        private void AddDownloadFilePage_OnDownloadClick(object? sender, EventArgs e)
+        {
+            
+        }
     }
 }
